move watches to up since it's only used there
